modularized parts of search use case
