+++ app/src/main/java/com/example/metmuseum/gallery/Extensions.kt
@@ -1,11 +1,8 @@
 package com.example.metmuseum.gallery
 
-import com.example.met_api.model.MetCollectionObject
-import com.example.met_api.model.MetObjectId
-
-fun com.example.met_api.model.MetCollectionObject.toSearchModel(): SearchModel {
+fun com.example.met_api.model.MetCollectionObject.toSearchModel(): com.example.entities.SearchModel {
     val idList = objectIds.map { com.example.met_api.model.MetObjectId(id = it) }
-    return SearchModel(
+    return com.example.entities.SearchModel(
         idList
     )
 }
