+++ build.gradle
@@ -15,4 +15,5 @@ plugins {
     id 'org.jetbrains.kotlin.android' version '1.8.0' apply false
     // Hilt
     id 'com.google.dagger.hilt.android' version '2.44' apply false
+    id 'org.jetbrains.kotlin.jvm' version '1.8.20' apply false
 }
